Move shared test methods into a helper file
We'll be able to re-use some fo tehse when we start building a test
suite for single_page_unpublishing_message_processor shortly.
