+++ spec/models/message_processor_spec.rb
@@ -1,6 +1,8 @@
 require "spec_helper"
 
 RSpec.describe MessageProcessor do
+  include MessageProcessorHelpers
+
   let(:logger) { double(:logger, info: nil) }
 
   let(:processor) do
@@ -15,18 +17,6 @@ RSpec.describe MessageProcessor do
       retry: nil,
       payload: "title['test']",
     )
-  end
-
-  def message_acknowledged
-    expect(message).to have_received(:ack)
-  end
-
-  def message_rejected
-    expect(message).to have_received(:discard)
-  end
-
-  def message_requeued
-    expect(message).to have_received(:retry)
   end
 
   describe "#process" do
